requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
requests==2.32.3
python-dotenv==1.0.1
gspread==6.1.3
oauth2client==4.1.3 